\name{Index.Econometrics}
\docType{documentation}
\alias{Index.Econometrics}
\title{Econometric fields}
\description{
\describe{
\item binomial model 
 \itemize{
\item \code{\link{LfpSwiss}} : Labor Force Participation in Swizerland  
\item \code{\link{Train}} : Stated Preferences for Train Traveling  
}
\item censored or truncated model 
 \itemize{
\item \code{\link{HI}} : Health Insurance and Hours Worked By Wives  
\item \code{\link{Mofa}} : International Expansion of U.s. Mofa's (majority--owned Foreign Affiliates in Fire (finance, Insurance and Real Estate)  
\item \code{\link{Workinghours}} : Wife Working Hours  
}
\item count data 
 \itemize{
\item \code{\link{PatentsRD}} : Patents, R\&d and Technological Spillovers for a Panel of Firms  
\item \code{\link{Somerville}} : Visits to Lake Somerville  
}
\item duration model 
 \itemize{
\item \code{\link{Oil}} : Oil Investment  
\item \code{\link{Unemployment}} : Unemployment Duration  
}
\item multinomial model 
 \itemize{
\item \code{\link{Car}} : Stated Preferences for Car Choice  
\item \code{\link{Catsup}} : Choice of Brand for Catsup  
\item \code{\link{Cracker}} : Choice of Brand for Crakers  
\item \code{\link{HC}} : Heating and Cooling System Choice in Newly Built Houses in California  
\item \code{\link{Heating}} : Heating System Choice in California Houses  
\item \code{\link{Ketchup}} : Choice of Brand for Ketchup  
\item \code{\link{Mode}} : Mode Choice  
\item \code{\link{Tuna}} : Choice of Brand for Tuna  
\item \code{\link{Yogurt}} : Choice of Brand for Yogurts  
}
\item ordered model 
 \itemize{
\item \code{\link{Kakadu}} : Willingness to Pay for the Preservation of the Kakadu National Park  
\item \code{\link{Mathlevel}} : Level of Calculus Attained for Students Taking Advanced Micro--economics  
}
\item system of equations 
 \itemize{
\item \code{\link{BudgetItaly}} : Budget Shares for Italian Households  
\item \code{\link{BudgetUK}} : Budget Shares of British Households  
\item \code{\link{University}} : Provision of University Teaching and Research  
}
}

}
\keyword{documentation}
