\name{Index.Economics}
\docType{documentation}
\alias{Index.Economics}
\title{Economic fields}
\description{
\describe{
\item consumer behaviour 
 \itemize{
\item \code{\link{Car}} : Stated Preferences for Car Choice  
\item \code{\link{HC}} : Heating and Cooling System Choice in Newly Built Houses in California  
\item \code{\link{Heating}} : Heating System Choice in California Houses  
\item \code{\link{Mode}} : Mode Choice  
\item \code{\link{Somerville}} : Visits to Lake Somerville  
\item \code{\link{Train}} : Stated Preferences for Train Traveling  
}
\item contingent valuation 
 \itemize{
\item \code{\link{Kakadu}} : Willingness to Pay for the Preservation of the Kakadu National Park  
}
\item hedonic prices 
 \itemize{
\item \code{\link{Computers}} : Prices of Personal Computers  
}
\item household budgets 
 \itemize{
\item \code{\link{BudgetFood}} : Budget Share of Food for Spanish Households  
\item \code{\link{BudgetItaly}} : Budget Shares for Italian Households  
\item \code{\link{BudgetUK}} : Budget Shares of British Households  
}
\item labor economics 
 \itemize{
\item \code{\link{HI}} : Health Insurance and Hours Worked By Wives  
\item \code{\link{LfpSwiss}} : Labor Force Participation in Swizerland  
\item \code{\link{Unemployment}} : Unemployment Duration  
\item \code{\link{Workinghours}} : Wife Working Hours  
}
\item marketing 
 \itemize{
\item \code{\link{Catsup}} : Choice of Brand for Catsup  
\item \code{\link{Cracker}} : Choice of Brand for Crakers  
\item \code{\link{Ketchup}} : Choice of Brand for Ketchup  
\item \code{\link{Tuna}} : Choice of Brand for Tuna  
\item \code{\link{Yogurt}} : Choice of Brand for Yogurts  
}
\item producer behavior 
 \itemize{
\item \code{\link{Mofa}} : International Expansion of U.s. Mofa's (majority--owned Foreign Affiliates in Fire (finance, Insurance and Real Estate)  
\item \code{\link{Oil}} : Oil Investment  
\item \code{\link{PatentsRD}} : Patents, R\&d and Technological Spillovers for a Panel of Firms  
\item \code{\link{University}} : Provision of University Teaching and Research  
}
\item socio--economics 
 \itemize{
\item \code{\link{Mathlevel}} : Level of Calculus Attained for Students Taking Advanced Micro--economics  
}
}

}
\keyword{documentation}
