\name{LfpSwiss}
\docType{data}
\alias{LfpSwiss}
\title{Labor Force Participation in Swizerland }
\description{
  a cross-section from 1981

 \emph{number of observations} :  872

 \emph{observation} :  individuals

   \emph{country} :  Swizerland

}
\usage{data(LfpSwiss)}
\format{A dataframe containing :
  \describe{
    \item{lfp}{   Labour force participation ?}
    \item{lnnlinc}{ Log of nonlabour income}
    \item{age}{ Age in years divided by 10}
    \item{educ}{years of formal education}
    \item{nyc}{ Number of young children (younger than 7)}
    \item{noc}{ Number of older children}
    \item{foreign}{not Swiss ?}
  }
}
\source{
    Gerfin, Michael (1996) \dQuote{Parapetric and semi--parametric estimation of the binary responsemodel of labour market participation}, \emph{Journal of Applied Econometrics}, \bold{11(3)}, may--june, 321--339.

}
\references{
      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}
\seealso{\code{\link{Index.References}}, \code{\link{Index.Economics}}, \code{\link{Index.Econometrics}}}}

\keyword{datasets}
